Late load channel endpoint
This moves all the API endpoints for channels into a single module
and loads that module late, when the service starts. There seems
to be some sort of dependency loop that breaks the oslo.config
config loading from an externally-defined file.
