+++ pyproject.toml
@@ -25,6 +25,8 @@ keystonemiddleware = "^7.0.1"
 "oslo.reports" = "^1.30.0"
 futurist = "^1.9.0"
 "oslo.rootwrap" = "^5.16.1"
+python-neutronclient = "^7.6.0"
+jsonschema = "^3.2.0"
 
 [tool.poetry.dev-dependencies]
 black = "^20.8b1"
