+++ tunelo/api/__init__.py
@@ -11,8 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from . import create_channel
-from . import destroy_channel
-from . import get_channel
-from . import list_channels
-from . import update_channel
